requirements.txt
boto3==1.28.20
botocore==1.31.20
jmespath==1.0.1
nose==1.3.7
psycopg2-binary==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
cryptography==3.4.7